Adding Order Authorization validate

+++ server/src/api/application/use_cases/order/deleteById.js
@@ -1,4 +1,11 @@
-const deleteById = (dbRepository, idOrder) => {
-  return dbRepository.deleteOrder(idOrder);
+const AuthorizationError = require('../../../middlewares/exceptions/AuthorizationError');
+
+const deleteById = (dbRepository, idOrder, idUser) => {
+  return dbRepository.findById(idOrder).then((result) => {
+    if (result.dataValues.idUser !== idUser) {
+      throw new AuthorizationError('You cannot access this resource');
+    }
+    return dbRepository.deleteOrder(idOrder);
+  });
 };
 module.exports = deleteById;
